feat: expose stats schema on Snapshot (#2128)
# Description

The new table config `delta.dataSkippingStatsColumns` changes the logic
which columns should be considered when computing file statistics. Here
we expose (and use) a new method on `Snapshot` which reflects the
updated logic.

This PR does not contain any tests for the new code paths, but I'll test
this against the reference implementation once the new checkpoint writer
lands.